+++ loop/sfm/logreg/utils/__init__.py
@@ -1,9 +1,0 @@
-# Make utils a Python package
-
-from loop.sfm.logreg.utils.breakout_regressor import (
-    build_sample_dataset_for_breakout_regressor
-)
-
-__all__ = [
-    'build_sample_dataset_for_breakout_regressor'
-]